Refactor: Move disabling delay to .retry
