+++ lib/client.js
@@ -2,7 +2,7 @@
 
 const compose = require('koa-compose');
 const context = require('./context');
-const delayBy = require('./delayedPromise');
+const rejectedPromise = require('./rejectedPromise');
 const packageInfo = require('../package');
 const bind = require('./bind');
 
@@ -338,9 +338,6 @@ class HttpTransport {
     const currentContext = this._ctx;
     currentContext.retries = this._retries;
     currentContext.retryDelay = this._retryDelay >= 0 ? this._retryDelay : RETRY_DELAY;
-    if (currentContext.retries === 0) {
-      currentContext.retryDelay = 0;
-    }
     this._initContext();
 
     return retry(this._executeRequest, currentContext).then((ctx) => ctx.res);
@@ -393,7 +390,13 @@ function retry(fn, ctx) {
 
   function attempt(i) {
     return request()
-      .catch(delayBy(ctx.retryDelay))
+      .catch((err) => {
+        if (ctx.retries > 0) {
+          const delayBy = rejectedPromise(ctx.retryDelay);
+          return delayBy(err);
+        }
+        throw err;
+      })
       .catch((err) => {
         if (i < ctx.retries && isCriticalError(err)) {
           ctx.res.retries.push(toRetry(err));
