+++ lib/rejectedPromise.js
